Add implementation of PaX features relevant to userland exploitation: PAGEEXEC MPROTECT ASLR (RANDUSTACK/RANDMMAP)
The implementation is specific to ARMv6+ so several simplifications have been
made for code clarity:
* EMUTRAMP and all other arch-specific config options to PaX have been removed
* Non-ARM-specific code related to the above features have been removed.  This includes
  the code in mm/mmap.c to override upstream ALSR in the top-down and bottom-up
  allocators for archs (unlike ARM) that don't provide their own arch-specific allocators.
* The legacy EI and PT PaX binary markings have been eliminated
* The bits of entropy for mmap/stack have been fixed to 16 since we don't need to worry
  about old ARM processors (which lack XN as well)

All non-trivial blocks of code have been commented to clearly define their purpose.
PaX-specific Kconfig options have been split out into security/Kconfig.PaX
Code to force on user xattrs for filesystems that support it has been included.
Code to support user.pax.flags on tmpfs has also been added.

This implementation is described by the following documents:
http://pax.grsecurity.net/docs/pageexec.txt
http://pax.grsecurity.net/docs/mprotect.txt
http://pax.grsecurity.net/docs/randustack.txt
http://pax.grsecurity.net/docs/randmmap.txt
http://pax.grsecurity.net/docs/aslr.txt

with the additional note that PROT_NONE mappings have been created to fill
the gap between the end of the main executable and the start of the brk-managed
heap as well as between the "top" of the randomized stack and the highest userland
address (the former to prevent .data overflows into the heap, the latter to aid
in preserving ASLR behavior under address space exhaustion).  Code exists in
this patch (and is commented as such) to prevent VM accounting for these PROT_NONE
mappings.

Compiled/tested against the current tree.