Implement a number of measures mostly provided under CONFIG_GRKERNSEC_PROC_MEMMAP to further harden PaX's ASLR implementation against information leakage.
The general protection model mimics that provided by grsecurity in that we're
concerned not only with leaks from processes of other users, but even from processes
of the same user, where you have a scenario where one process has heavy access
control forced on it by SELinux or the like and another process running under
the same UID has less access control enforcement.  The easiest way to resolve
this while maintaining compatibility with important cases like glibc's %n
protection (which reads /proc/self/maps to see if a use of %n is from a writable
format string) is to deny visibility into the ASLR layout to any process with
ASLR enabled, except for self-inspection.

Several /proc entries are sanitizied using the above model, including
maps, smaps, numa, and stat.  It's also noteworthy that support for
/proc/pid/syscall is disabled completely because it's impossible for
it to be fixed to not leak ASLR-defeating information via its reported
syscall arguments.

Upstream flaws in move_pages() and migrate_pages() are fixed which allow
ASLR layout discovery by repeated syscall probing against suid/sgid apps.

A per-task ID is created and updated on each execve to allow for quick
determination whether the task that opened a particular sensitive /proc
entry is also the one that is trying to read from or write to it.  This
applies to /proc/pid/(mem, environ, maps, smaps, numa, and others)

Writing to /proc/pid/mem is explicitly disallowed as it gives a
filesystem-only related bug the ability to corrupt the memory
of the current process.

Various methods of effectively disabling ASLR for suid/sgid binaries
by lowering or raising resource limits have also been thwarted in
this patch.

Every snippet of code in this commit is commented fully to explain
its purpose.

Compiled and tested against the current tree.