Implement invisible sigreturn stub page
Upstream has a vulnerability in their sigreturn page implementation:
the code comments say some poisoning is added to the page to fill
in the rest of the page outside where the sigreturn stubs are located,
but no actual poisoning is done, leaving the bytes as 0 which act
as a simple no-op.  The intra-page offset they implement is also only
computed once and shared among all processes.

This patch also disables VDSO support in the kernel, as it wants to
piggy-back off the sigpage allocation.

This patch improves the ASLR implementation on ARM, as sigreturn is
a very useful for exploitation (it allows one to set arbitrary
register context).